Liquibase

+++ src/main/java/ru/pnzgu/restauran/dto/SostavProdDTO.java
@@ -1,5 +1,6 @@
 package ru.pnzgu.restauran.dto;
 
+import liquibase.pro.packaged.D;
 import lombok.*;
 
 import java.math.BigDecimal;
@@ -12,6 +13,6 @@ import java.math.BigDecimal;
 public class SostavProdDTO extends DtoParent {
     private MenuDTO bludo;
     private Long quantity = 0L;
-    private BigDecimal summa = BigDecimal.valueOf(0);
+    private Double summa = (double) 0;
     private ProdazaDTO prodaza;
 }
